finish timeslot

+++ YogaManagement/YogaManagement.Application/Controllers/SchedulesController.cs
@@ -1,10 +1,13 @@
 ﻿using AutoMapper;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.OData.Deltas;
 using Microsoft.AspNetCore.OData.Query;
 using Microsoft.AspNetCore.OData.Routing.Controllers;
+using Microsoft.EntityFrameworkCore.Metadata.Internal;
 using YogaManagement.Application.Utilities;
 using YogaManagement.Business.Repositories;
 using YogaManagement.Contracts.TimeSlot;
+using YogaManagement.Domain.Enums;
 using YogaManagement.Domain.Models;
 
 namespace YogaManagement.Application.Controllers;
@@ -12,11 +15,15 @@ public class SchedulesController : ODataController
 {
     private readonly IMapper _mapper;
     private readonly ScheduleRepository _scheduleRepo;
+    private readonly YogaClassRepository _ygClassRepo;
 
-    public SchedulesController(ScheduleRepository Repo, IMapper mapper)
+    public SchedulesController(ScheduleRepository Repo,
+        YogaClassRepository ygClassRepo,
+        IMapper mapper)
     {
         _mapper = mapper;
         _scheduleRepo = Repo;
+        _ygClassRepo = ygClassRepo;
     }
     //get schedules of a class
     [EnableQuery]
@@ -27,8 +34,19 @@ public class SchedulesController : ODataController
 
     public async Task<IActionResult> Post([FromBody] ScheduleDTO createRequest)
     {
+        var existClass = await _ygClassRepo.Get(createRequest.YogaClassId);
+        if (existClass == null)
+        {
+            return NotFound();
+        }
+
         try
         {
+            if (existClass.YogaClassStatus == YogaClassStatus.Active && existClass.Course.EnddDate < DateTime.Today)
+            {
+                throw new Exception("Cannot delete ongoing class");
+            }
+
             ModelState.ValidateRequest();
             var newSchedule = _mapper.Map<Schedule>(createRequest);
             await _scheduleRepo.CreateAsync(newSchedule);
@@ -43,6 +61,12 @@ public class SchedulesController : ODataController
     // delete 1 slot 
     public async Task<IActionResult> Delete([FromRoute] int keyTimeSlotId, [FromRoute] int keyYogaClassId)
     {
+        var existClass = await _ygClassRepo.Get(keyYogaClassId);
+        if (existClass == null)
+        {
+            return NotFound();
+        }
+
         var existSchedule = _scheduleRepo.GetSchedule(keyTimeSlotId, keyYogaClassId);
         if (existSchedule == null)
         {
@@ -50,6 +74,11 @@ public class SchedulesController : ODataController
         }
         try
         {
+            if (existClass.YogaClassStatus == YogaClassStatus.Active && existClass.Course.EnddDate < DateTime.Today)
+            {
+                throw new Exception("Cannot delete ongoing class");
+            }
+
             await _scheduleRepo.DeleteAsync(existSchedule);
             return NoContent();
         }
@@ -63,6 +92,12 @@ public class SchedulesController : ODataController
     [HttpDelete("odata/Schedules/yogaClassId")]
     public async Task<IActionResult> Delete(int keyYogaClassId)
     {
+        var existClass = await _ygClassRepo.Get(keyYogaClassId);
+        if (existClass == null)
+        {
+            return NotFound();
+        }
+
         var listSchedule = _scheduleRepo.GetScheduleOfAClass(keyYogaClassId);
         if (listSchedule == null)
         {
@@ -70,6 +105,11 @@ public class SchedulesController : ODataController
         }
         try
         {
+            if (existClass.YogaClassStatus == YogaClassStatus.Active && existClass.Course.EnddDate < DateTime.Today)
+            {
+                throw new Exception("Cannot delete ongoing class");
+            }
+
             foreach (var schedule in listSchedule)
             {
                 await _scheduleRepo.DeleteAsync(schedule);
